Separate d3 logic to component map-chart-view
