+++ components/map/map-chart-view.css
@@ -1,0 +1,35 @@
+#map {
+  .drifter {
+    fill: blue;
+    opacity: 0.3;
+    cursor: pointer;
+
+
+  }
+
+  .drifter:hover {
+
+    fill: green;
+    opacity: 1;
+
+  }
+}
+
+
+
+
+#map .drifter.selected {
+  fill: red;
+  opacity: 1;
+}
+
+
+.lineConnect {
+  /* fill: magenta; */
+  /* stroke: black; */
+  opacity: 1;
+  fill: none;
+  stroke: green;
+  opacity: 0.5;
+  stroke-width: 3;
+}
